requirements.txt
Flask==1.1.2
Flask-SQLALchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==2.7.0
redis==3.5.3
oauthlib==3.1.0
pyOpenSSL==20.0.1
pymysql==1.0.2
requests==2.25.1
pyjwt==2.0.1
gunicorn==20.0.4
gevent==21.1.2
